+++ wp-content/themes/hestia/assets/bootstrap/js/bootstrap.js
@@ -696,7 +696,7 @@ if (typeof jQuery === 'undefined') {
 					this.options.placement.call( this, $tip[0], this.$element[0] ) :
 					this.options.placement
 
-					var autoToken = / \s ?auto ?\s ? /i
+					var autoToken = / \s ? auto ? \s ? / i
 					var autoPlace = autoToken.test( placement )
 					if (autoPlace) {
 						placement = placement.replace( autoToken, '' ) || 'top'