+++ wp-content/themes/hestia/template-parts/content-page.php
@@ -20,40 +20,45 @@ $args         = array(
 $class_to_add = hestia_get_content_classes( $hestia_page_sidebar_layout, 'sidebar-1', $args );
 ?>
 
-<article id="post-<?php the_ID(); ?>" class="section section-text">
-	<div class="row">
-		<?php
-		if ( $hestia_page_sidebar_layout === 'sidebar-left' ) {
-			hestia_get_sidebar();
-		}
-		?>
-		<div class="<?php echo esc_attr( $class_to_add ); ?>">
-			<?php the_content(); ?>
+	<article id="post-<?php the_ID(); ?>" class="section section-text">
+		<div class="row">
+			<?php
+			if ( $hestia_page_sidebar_layout === 'sidebar-left' ) {
+				hestia_get_sidebar();
+			}
+			?>
+			<div class="<?php echo esc_attr( $class_to_add ); ?>">
+				<?php the_content(); ?>
+			</div>
+			<?php
+			if ( $hestia_page_sidebar_layout === 'sidebar-right' ) {
+				hestia_get_sidebar();
+			}
+			?>
 		</div>
-		<?php
-		if ( $hestia_page_sidebar_layout === 'sidebar-right' ) {
-			hestia_get_sidebar();
-		}
-		?>
-	</div>
-</article>
-<div class="section section-blog-info">
-	<div class="row">
-		<div class="col-md-8 col-md-offset-2">
-			<div class="row">
-				<div class="col-md-12">
-					<?php
-					hestia_wp_link_pages(
-						array(
-							'before'      => '<div class="text-center"> <ul class="nav pagination pagination-primary">',
-							'after'       => '</ul> </div>',
-							'link_before' => '<li>',
-							'link_after'  => '</li>',
-						)
-					);
-					?>
+	</article>
+<?php
+if ( is_paged() ) {
+	?>
+	<div class="section section-blog-info">
+		<div class="row">
+			<div class="col-md-8 col-md-offset-2">
+				<div class="row">
+					<div class="col-md-12">
+						<?php
+						hestia_wp_link_pages(
+							array(
+								'before'      => '<div class="text-center"> <ul class="nav pagination pagination-primary">',
+								'after'       => '</ul> </div>',
+								'link_before' => '<li>',
+								'link_after'  => '</li>',
+							)
+						);
+						?>
+					</div>
 				</div>
 			</div>
 		</div>
 	</div>
-</div>
+	<?php
+}